+++ inst/include/procedures/riemfunc_nearest.h
@@ -14,8 +14,12 @@ inline arma::mat riemfunc_nearest(arma::mat x, std::string name){
     return(sphere_nearest(x));
   } else if (name=="spd"){
     return(spd_nearest(x));
-  }else {
-    Rcpp::Rcout << "RiemBase::riemfunc_nearest : " << name << "is not yet implemented." << std::endl;
+  } else if (name=="stiefel"){
+    return(stiefel_nearest(x));
+  } else if (name=="grassmann"){
+    return(grassmann_nearest(x));
+  } else {
+    Rcpp::Rcout << "RiemBase::riemfunc_nearest : " << name << " is not yet implemented." << std::endl;
     Rcpp::stop("");
   }
 }